--- pages/climate-modeling.vue
@@ -1,255 +1,0 @@
-<template>
-  <section class="section">
-    <div class="content is-size-5">
-      <h2 id="why-decision-support" class="title is-3">
-        Why we need easy&ndash;to&ndash;use decision&ndash;support tools for
-        climate adaptation
-      </h2>
-      <ul class="toc">
-        <li>
-          <nuxt-link :to="{ hash: 'why-decision-support' }"
-            >Why do we need this tool?</nuxt-link
-          >
-        </li>
-        <li>
-          <nuxt-link :to="{ hash: 'arctic-change' }"
-            >What&rsquo;s changing in the Arctic?</nuxt-link
-          >
-        </li>
-        <li>
-          <nuxt-link :to="{ hash: 'models-projections-uncertainty' }"
-            >Models, projections, and uncertainty</nuxt-link
-          >
-        </li>
-        <li>
-          <nuxt-link :to="{ hash: 'how-projections-work' }"
-            >How are climate projections made?</nuxt-link
-          >
-        </li>
-        <li>
-          <nuxt-link :to="{ hash: 'best-practices' }"
-            >Best practices for producing and using climate
-            projections</nuxt-link
-          >
-        </li>
-        <li>
-          <nuxt-link :to="{ hash: 'real-compare-projection' }"
-            >How might the real climate conditions compare to
-            projections?</nuxt-link
-          >
-        </li>
-      </ul>
-
-      <p>
-        Assessments of climate&ndash;change effects in the Alaska are hindered
-        by a lack of decision&ndash;support tools that can visualize possible
-        future landscapes under different scenarios.
-      </p>
-      <p>
-        Model output products are publicly available, but often in formats
-        requiring considerable expertise to access and process. Repackaging
-        these outputs as climate reports moves relevant climate information out
-        of storage and into decision&ndash;making spaces.
-      </p>
-
-      <h2 id="arctic-change" class="title is-3">
-        What&rsquo;s changing in the Arctic?
-      </h2>
-      <p>
-        Using climate data to envision how changes may progress in the
-        near&ndash;term and more distant future is crucial for engineers,
-        planners, and communities.
-      </p>
-      <ul>
-        <li>
-          <strong>Warming temperatures</strong> are driving permafrost thaw and
-          changes in heating degree days, cooling degree days, sea and river
-          ice, glaciation, and more.
-        </li>
-        <li>
-          Thawing of <strong>permafrost</strong> alters surface drainage and
-          ground stability.
-        </li>
-        <li>
-          Changes in <strong>precipitation</strong>, including shifts from snow
-          to rain and altered frequency of extreme events can affect roads and
-          structures.
-        </li>
-        <li>
-          Variables not directly linked to climate, such as daylight and tides,
-          may be experienced differently in an altered climate.
-        </li>
-      </ul>
-      <p>
-        These connections are complex, and the uncertainty associated with them
-        is high. Nonetheless, by studying and modeling each factor and
-        connecting model outputs when possible, we can produce a range of
-        projections that explore possible changes locally, regionally, or
-        landscape-wide across the North.
-      </p>
-      <h2 id="models-projections-uncertainty" class="title is-3">
-        Models, projections, uncertainties
-      </h2>
-      <h3 class="subtitle is-4">
-        Climate models help us imagine possible climate futures.
-      </h3>
-      <p>
-        We use weather forecasts for short-term planning. Climate projections
-        can be used for long&ndash;term planning&mdash;but they are not the same
-        as forecasts. Weather varies day to day, whereas climate refers to the
-        average or typical conditions over much longer time periods. However,
-        there are uncertainties based on model limitations and unknown future
-        human behavior that make long-range forecasting very different from
-        predicting tomorrow's weather.
-      </p>
-      <p>
-        Climate projections look much further into the future than weather
-        forecasts. They address uncertainties by focusing on the range of future
-        conditions that would likely occur, given what we currently know about
-        the climate system and how it will change in response to changes in the
-        factors that affect it.
-      </p>
-
-      <h3 id="how-projections-work" class="subtitle is-4">
-        How do we make these projections? We use climate models.
-      </h3>
-      <p>
-        Climate models are simplified versions of reality that try to explain
-        climatic processes with just the most necessary parts of the system.
-        Their usefulness is evident when we compare observed historical climate
-        and simulated data&mdash;the models capture the most important climate
-        patterns.
-      </p>
-      <p>
-        Climate models use data to calculate how the global climate varies.
-        These data include:
-      </p>
-      <ul>
-        <li>initial climate conditions</li>
-        <li>
-          &ldquo;forcings&rdquo; such as atmospheric greenhouse gas, solar and
-          volcanic variability
-        </li>
-        <li>ocean and atmosphere variability</li>
-        <li>land surface conditions</li>
-        <li>feedbacks such as the carbon cycle and the water cycle</li>
-      </ul>
-      <p>
-        The end product is a simulation of future climate. Because the end
-        product is based on statistical probabilities, the data are most
-        reliable when averaged across time or space, such as the projected
-        average of 30 years of winter precipitation for your community, or the
-        likely hottest temperature that might occur on the North Slope.
-      </p>
-
-      <h3 class="subtitle is-4">
-        Climate models aren&rsquo;t perfect, though.
-      </h3>
-      <p>
-        Between now and about 20&ndash;30 years from now, current climate change
-        is the best predictor of the rate of change, but year to year
-        variability is the largest source of uncertainty. Although long-term
-        climate change shows clear trends, those can be masked by natural ups
-        and downs in the short term. Climate models do simulate this kind of
-        variability, but they cannot predict it precisely.
-      </p>
-
-      <h3 id="best-practices" class="subtitle is-4">
-        Best practices for making projections
-      </h3>
-      <ul>
-        <li>
-          <strong>Use multiple decades.</strong> Averages over 20&ndash;30 years
-          are more resistant to transient variability in climate models and to
-          natural variations in regional climate. Compare a future (like
-          2030&ndash;2069) to a historical reference frame (like
-          1970&ndash;1999), and keep in mind that the later the historical
-          reference frame, the more climate change is already in it!
-        </li>
-        <li>
-          <strong>Use multiple climate models.</strong> Picking one model is not
-          good practice because all the models are at least plausible, if not
-          equally likely. Use several separate models if the full range of
-          possibilities is important to your work, or use the average of
-          multiple models if you are more interested in the most likely outputs.
-          This is especially critical between now and about the 2050s or 2060s.
-        </li>
-        <li>
-          <strong>Use multiple emission scenarios.</strong> Given the
-          uncertainty around future human behavior, you should pick at least two
-          scenarios that bracket the likely range, unless you are only
-          interested in looking at the &ldquo;best case&rdquo; or &ldquo;worst
-          case&rdquo;. RCP 4.5 and RCP 8.5 are good choices.. This is especially
-          critical after about 2050.
-        </li>
-        <li>
-          <strong
-            >Look at medium&ndash;term and longer&ndash;term futures.</strong
-          >
-          A comprehensive assessment would consider a historical, a mid-21st
-          century future, and a late&ndash;21<sup>st</sup> century future. The
-          two futures should have a high, low and middle range each, possibly
-          with multiple models and multiple emission scenarios in each future
-          window.
-        </li>
-        <li>
-          <strong>Don&rsquo;t make your assessment area too small.</strong>
-          The more local your assessment, the more likely it is that local
-          factors like elevation, vegetation, etc. and the process used to
-          downscale the climate model to local resolution contribute to the
-          uncertainty. Larger areas are probably more resistant to this local
-          variation, so a watershed, planning unit, or responses across several
-          of these are perhaps more useful to consider.
-        </li>
-        <li>
-          <strong
-            >For fire projections and post-fire vegetation, look at averages
-            across many model runs.</strong
-          >
-          The ALFRESCO fire and vegetation model cannot predict the precise
-          behavior of future fires, but it can simulate the likelihood that
-          fires will start and spread across broad landscapes across long
-          periods of time, causing shifts in the age and type of vegetation.
-          When looking at outputs, consider either a typical model run
-          (&ldquo;best replicate&rdquo;) or an average of many possible burning
-          scenarios. Outputs are helpful for assessing large&ndash;scale
-          long&ndash;term shifts, but are not meaningful at a pixel by pixel
-          level.
-        </li>
-      </ul>
-
-      <h3 id="real-compare-projection" class="subtitle is-4">
-        How will the &ldquo;real&rdquo; climate compare to any projection?
-      </h3>
-
-      <p>
-        The future climate we experience will not look exactly like any of these
-        projections, but it will look like a lot of them. There are a range of
-        future climates we may experience, and best practices are to plan for
-        the likely range of climates, impacts, and associated risks for the time
-        frame and region you&rsquo;re planning for.
-      </p>
-      <p>
-        <strong
-          >Projections are always improving incrementally. Don&rsquo;t wait for
-          a better projection&mdash;you&rsquo;ll always be waiting and the costs
-          of waiting will increase. In general, plan for the range of historical
-          variability plus the range of climates described by a less warm (none
-          show cooling!) climate model under a lower emissions scenario.</strong
-        >
-      </p>
-    </div>
-  </section>
-</template>
-
-<style lang="scss" scoped>
-p:last-of-type {
-  margin-bottom: 2.5rem;
-}
-</style>
-
-<script>
-export default {
-  layout: 'content',
-}
-</script>
